2015.9.18：保持路径概率m_NotSelectBestRoutePathRate修改为double

+++ ns-3/src/pmtmgmp/model/my11s/pmtmgmp-rtable.h
@@ -264,7 +264,7 @@ namespace ns3 {
 			Time m_AcceptInformaitonDelay;
 
 			////非当前使用路径最优路径比较倍率
-			uint8_t m_NotSelectBestRoutePathRate;
+			double m_NotSelectBestRoutePathRate;
 		};
 		/*************************
 		* PmtmgmpRouteTable
